requirements.txt
# 项目依赖包列表
# 注意：安装顺序很重要，避免版本冲突和覆盖

# 音频处理依赖
ffmpeg-python>=0.2.0

# 语音识别核心依赖
faster-whisper>=0.6.0

# 中文繁简转换
zhconv>=1.4.0

# PyTorch (需要CUDA支持)
torch>=1.10.0+cu113